@@ -1,4 +1,4 @@
-// frontend/lib/screens/welcome_screen.dart
+// Path: frontend/lib/screens/welcome_screen.dart
 
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -15,16 +15,17 @@ class _WelcomeScreenState extends State<WelcomeScreen> {
   @override
   void initState() {
     super.initState();
-    _checkLoginStatus();
+    // Use addPostFrameCallback to ensure the context is available
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      _checkLoginStatus();
+    });
   }
 
   Future<void> _checkLoginStatus() async {
-    // Use the AuthService from the Provider, don't create a new one.
     final authService = Provider.of<AuthService>(context, listen: false);
     await authService.tryAutoLogin();
-    if (authService.isAuthenticated) {
-      // If authenticated, navigate to the main flow
-      Navigator.of(context).pushReplacementNamed('/goal');
+    if (mounted && authService.isAuthenticated) {
+      Navigator.of(context).pushReplacementNamed('/creation_hub');
     }
   }
 
@@ -32,28 +33,54 @@ class _WelcomeScreenState extends State<WelcomeScreen> {
   Widget build(BuildContext context) {
     return Scaffold(
       body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'Artisan AI',
-              style: Theme.of(context).textTheme.headlineLarge,
-            ),
-            const SizedBox(height: 48),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.pushNamed(context, '/login');
-              },
-              child: const Text('Login'),
-            ),
-            const SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.pushNamed(context, '/register');
-              },
-              child: const Text('Register'),
-            ),
-          ],
+        child: Padding(
+          padding: const EdgeInsets.all(24.0),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: <Widget>[
+              const Icon(
+                Icons.hub_outlined, // Icon suggesting connection/intelligence
+                size: 80,
+                color: Color(0xFF81B2D9),
+              ),
+              const SizedBox(height: 24),
+              Text(
+                'Synaptiq.ai',
+                textAlign: TextAlign.center,
+                style: Theme.of(context).textTheme.displaySmall?.copyWith(
+                      color: Colors.white,
+                    ),
+              ),
+              const SizedBox(height: 12),
+              Text(
+                'Your Intelligent Prompt Co-Pilot',
+                textAlign: TextAlign.center,
+                style: Theme.of(context).textTheme.headlineSmall?.copyWith(
+                      color: Colors.white70,
+                    ),
+              ),
+              const SizedBox(height: 64),
+              ElevatedButton(
+                onPressed: () {
+                  Navigator.pushNamed(context, '/login');
+                },
+                child: const Text('Login'),
+              ),
+              const SizedBox(height: 16),
+              OutlinedButton(
+                onPressed: () {
+                  Navigator.pushNamed(context, '/register');
+                },
+                style: OutlinedButton.styleFrom(
+                  side: BorderSide(color: Theme.of(context).colorScheme.primary),
+                  padding: const EdgeInsets.symmetric(horizontal: 30, vertical: 15),
+                  textStyle: Theme.of(context).elevatedButtonTheme.style?.textStyle,
+                ),
+                child: const Text('Register'),
+              ),
+            ],
+          ),
         ),
       ),
     );
